requirements.txt
hydra-core~=1.3.2
omegaconf~=2.3.0
requests~=2.32.4
tqdm~=4.67.1
pandas~=2.3.1
pyarrow~=21.0.0
joblib~=1.5.1
scipy~=1.16.1
scikit-learn~=1.7.1
pytest~=8.4.1
numpy~=2.3.2